web_app/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
sentence-transformers
chromadb
pypdf
python-dotenv
